Clean up native code generation
