+++ src/tensora/desugar/__init__.py
@@ -1,0 +1,6 @@
+from .ast import Assignment, Integer, Float, Scalar, Tensor, Add, Multiply, Contract
+from .collect_lattices import collect_lattices
+from .desugar_expression import desugar_assignment
+from .id import Id
+from .to_identifiable import to_identifiable
+from .to_iteration_graph import to_iteration_graph
