+++ src/tensora/desugar/collect_lattices.py
@@ -3,8 +3,8 @@ __all__ = ["collect_lattices"]
 from functools import singledispatch
 from tensora.format.format import Format
 
-from tensora.iteration_graph.merge_lattice.merge_lattice import LatticeConjuction, LatticeDisjunction, LatticeLeaf, Lattice
 from . import ast
+from ..iteration_graph import LatticeConjuction, LatticeDisjunction, LatticeLeaf, Lattice
 from ..iteration_graph.identifiable_expression import ast as id
