+++ src/tensora/ir/__init__.py
@@ -1,0 +1,3 @@
+from .ast import Statement, Expression, Assignable, Variable, AttributeAccess, ArrayIndex, IntegerLiteral, FloatLiteral, BooleanLiteral, ModeLiteral, ArrayLiteral, Add, Subtract, Multiply, Equal, NotEqual, GreaterThan, GreaterThanOrEqual, LessThan, LessThanOrEqual, And, Or, FunctionCall, Max, Min, Address, BooleanToInteger, Allocate, ArrayAllocate, ArrayReallocate, Free, Declaration, Assignment, DeclarationAssignment, Block, Branch, Loop, Break, Return, FunctionDefinition
+from .builder import SourceBuilder
+from .peephole import peephole
