+++ src/tensora/iteration_graph/__init__.py
@@ -1,0 +1,3 @@
+from .iteration_graph_to_c_code import generate_c_code, KernelType
+from .problem import Problem
+from .merge_lattice import Lattice, LatticeLeaf, LatticeConjuction, LatticeDisjunction, IterationMode
