+++ src/tensora/iteration_graph/identifiable_expression/__init__.py
@@ -1,2 +1,7 @@
 from .ast import Assignment, Expression, Variable
+from .exhaust_tensors import exhaust_tensor
+from .index_dimension import index_dimension
+from .index_dimensions import index_dimensions
 from .tensor_leaf import TensorLeaf
+from .to_ir import to_ir
+from .variables import to_c_code
