+++ src/tensora/iteration_graph/merge_lattice/__init__.py
@@ -1,1 +1,1 @@
-from .merge_lattice import *
+from .merge_lattice import Lattice, LatticeLeaf, LatticeConjuction, LatticeDisjunction, IterationMode
