+++ tests/codegen/test_ast_to_c.py
@@ -80,12 +80,12 @@ single_lines = [
     (Declaration(Variable("x"), float), "double x"),
     (Declaration(Variable("x"), tensor), "taco_tensor_t x"),
     (Declaration(Variable("x"), hash_table), "hash_table_t x"),
-    (Declaration(Variable("x"), Pointer(float)), "double * restrict x"),
-    (Declaration(Variable("x"), Pointer(Pointer(integer))), "int32_t * restrict * restrict x"),
-    (Declaration(Variable("x"), Array(float)), "double[] x"),
-    (Declaration(Variable("x"), Array(Array(integer))), "int32_t[][] x"),
-    (Declaration(Variable("x"), FixedArray(mode, 3)), "taco_mode_t[3] x"),
-    (Declaration(Variable("x"), FixedArray(FixedArray(mode, 3), 2)), "taco_mode_t[3][2] x"),
+    (Declaration(Variable("x"), Pointer(float)), "double* restrict x"),
+    (Declaration(Variable("x"), Pointer(Pointer(integer))), "int32_t* restrict* restrict x"),
+    (Declaration(Variable("x"), Array(float)), "double x[]"),
+    (Declaration(Variable("x"), Array(Array(integer))), "int32_t x[][]"),
+    (Declaration(Variable("x"), FixedArray(mode, 3)), "taco_mode_t x[3]"),
+    (Declaration(Variable("x"), FixedArray(FixedArray(mode, 3), 2)), "taco_mode_t x[3][2]"),
     # Assignment
     (Assignment(Variable("x"), Variable("y")), "x = y"),
     (Assignment(Variable("x"), Add(Variable("x"), IntegerLiteral(1))), "x++"),
